backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
google-generativeai
sqlalchemy
numpy
